requirements.txt
beautifulsoup4==4.12.3
Flask==3.0.3
Flask_Cors==4.0.1
google_api_python_client==2.84.0
joblib==1.4.2
lxml==5.2.2
pandas==2.2.2
python-dotenv==1.0.1
pytube==15.0.0
Requests==2.32.3
gunicorn